# backend-spring/src/main/java/com/example/caseplatform/dto/CaseResponse.java
package com.example.caseplatform.dto;

import com.example.caseplatform.domain.enums.CaseStatus;
import java.time.Instant;
import java.util.UUID;

public class CaseResponse {

    private Long version;

    public Long getVersion() { return version; }
    public void setVersion(Long version) { this.version = version; }

    private UUID id;
    private String title;
    private String description;
    private CaseStatus status;
    private Instant createdAt;
    private Instant updatedAt;

    public UUID getId() {
        return id;
    }

    public void setId(UUID id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public CaseStatus getStatus() {
        return status;
    }

    public void setStatus(CaseStatus status) {
        this.status = status;
    }

    public Instant getCreatedAt() {
        return createdAt;
    }

    public void setCreatedAt(Instant createdAt) {
        this.createdAt = createdAt;
    }

    public Instant getUpdatedAt() {
        return updatedAt;
    }

    public void setUpdatedAt(Instant updatedAt) {
        this.updatedAt = updatedAt;
    }
}