# backend-spring/src/main/java/com/example/caseplatform/dto/OrderResponse.java
package com.example.caseplatform.dto;

import com.example.caseplatform.domain.enums.OrderStatus;
import java.time.Instant;
import java.util.UUID;

public class OrderResponse {

    private Long version;

    public Long getVersion() { return version; }
    public void setVersion(Long version) { this.version = version; }

    private UUID id;
    private UUID caseId;
    private String type;
    private OrderStatus status;
    private Instant createdAt;

    public UUID getId() {
        return id;
    }

    public void setId(UUID id) {
        this.id = id;
    }

    public UUID getCaseId() {
        return caseId;
    }

    public void setCaseId(UUID caseId) {
        this.caseId = caseId;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public OrderStatus getStatus() {
        return status;
    }

    public void setStatus(OrderStatus status) {
        this.status = status;
    }

    public Instant getCreatedAt() {
        return createdAt;
    }

    public void setCreatedAt(Instant createdAt) {
        this.createdAt = createdAt;
    }
}