# docs/openapi.json
{
  "components": {
    "schemas": {
      "AuditEventResponse": {
        "properties": {
          "actor": {
            "type": "string"
          },
          "caseId": {
            "format": "uuid",
            "type": "string"
          },
          "entityId": {
            "format": "uuid",
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "eventType": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {},
            "type": "object"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CaseCreateRequest": {
        "properties": {
          "description": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
      "CaseResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "CLOSED"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "CaseUpdateRequest": {
        "properties": {
          "description": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "CLOSED"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "pattern": "(?s).*\\S.*",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "type": "object"
      },
      "OrderCreateRequest": {
        "properties": {
          "caseId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "SENT",
              "COMPLETED",
              "FAILED"
            ],
            "type": "string"
          },
          "type": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "caseId",
          "type"
        ],
        "type": "object"
      },
      "OrderResponse": {
        "properties": {
          "caseId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "SENT",
              "COMPLETED",
              "FAILED"
            ],
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "OrderUpdateRequest": {
        "properties": {
          "status": {
            "enum": [
              "PENDING",
              "SENT",
              "COMPLETED",
              "FAILED"
            ],
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "type": "object"
      },
      "PageAuditEventResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PageCaseResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/CaseResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PageOrderResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PageableObject": {
        "properties": {
          "offset": {
            "format": "int64",
            "type": "integer"
          },
          "pageNumber": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "paged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SortObject": {
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Order Case Platform"
    },
    "description": "Case and order management service",
    "title": "Order Case Platform API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/cases": {
      "get": {
        "operationId": "getCases",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "CLOSED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCaseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "case-controller"
        ]
      },
      "post": {
        "operationId": "createCase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "case-controller"
        ]
      }
    },
    "/api/cases/{id}": {
      "delete": {
        "operationId": "deleteCase",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "case-controller"
        ]
      },
      "get": {
        "operationId": "getCase",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "case-controller"
        ]
      },
      "put": {
        "operationId": "updateCase",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "case-controller"
        ]
      }
    },
    "/api/cases/{id}/audit": {
      "get": {
        "operationId": "getCaseAudit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageAuditEventResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "case-controller"
        ]
      }
    },
    "/api/orders": {
      "get": {
        "operationId": "getOrders",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "caseId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageOrderResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "order-controller"
        ]
      },
      "post": {
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "order-controller"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "getOrder",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "order-controller"
        ]
      },
      "put": {
        "operationId": "updateOrder",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "order-controller"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}